+++ src/Loader/Optimized/mmdDataSerializer.ts
@@ -1,6 +1,4 @@
 export class MmdDataSerializer {
-    private static readonly _LittleEndian = true;
-
     private readonly _dataView: DataView;
     private readonly _encoder: TextEncoder;
     private _offset: number;
@@ -46,53 +44,53 @@ export class MmdDataSerializer {
     }
 
     public setUint16(value: number): void {
-        this._dataView.setUint16(this._offset, value, MmdDataSerializer._LittleEndian);
+        this._dataView.setUint16(this._offset, value, true);
         this._offset += 2;
     }
 
     public setUint16Array(values: ArrayLike<number>): void {
         const dataView = this._dataView;
         for (let i = 0; i < values.length; ++i) {
-            dataView.setUint16(this._offset, values[i], MmdDataSerializer._LittleEndian);
+            dataView.setUint16(this._offset, values[i], true);
             this._offset += 2;
         }
     }
 
     public setUint32(value: number): void {
-        this._dataView.setUint32(this._offset, value, MmdDataSerializer._LittleEndian);
+        this._dataView.setUint32(this._offset, value, true);
         this._offset += 4;
     }
 
     public setUint32Array(values: ArrayLike<number>): void {
         const dataView = this._dataView;
         for (let i = 0; i < values.length; ++i) {
-            dataView.setUint32(this._offset, values[i], MmdDataSerializer._LittleEndian);
+            dataView.setUint32(this._offset, values[i], true);
             this._offset += 4;
         }
     }
 
     public setInt32(value: number): void {
-        this._dataView.setInt32(this._offset, value, MmdDataSerializer._LittleEndian);
+        this._dataView.setInt32(this._offset, value, true);
         this._offset += 4;
     }
 
     public setInt32Array(values: ArrayLike<number>): void {
         const dataView = this._dataView;
         for (let i = 0; i < values.length; ++i) {
-            dataView.setInt32(this._offset, values[i], MmdDataSerializer._LittleEndian);
+            dataView.setInt32(this._offset, values[i], true);
             this._offset += 4;
         }
     }
 
     public setFloat32(value: number): void {
-        this._dataView.setFloat32(this._offset, value, MmdDataSerializer._LittleEndian);
+        this._dataView.setFloat32(this._offset, value, true);
         this._offset += 4;
     }
 
     public setFloat32Array(values: ArrayLike<number>): void {
         const dataView = this._dataView;
         for (let i = 0; i < values.length; ++i) {
-            dataView.setFloat32(this._offset, values[i], MmdDataSerializer._LittleEndian);
+            dataView.setFloat32(this._offset, values[i], true);
             this._offset += 4;
         }
     }
@@ -100,7 +98,7 @@ export class MmdDataSerializer {
     public setString(value: string): void {
         const dataView = this._dataView;
         const bytes = this._encoder.encode(value);
-        dataView.setUint32(this._offset, bytes.length, MmdDataSerializer._LittleEndian);
+        dataView.setUint32(this._offset, bytes.length, true);
         this._offset += 4;
         for (let i = 0; i < bytes.length; ++i) {
             dataView.setUint8(this._offset, bytes[i]);